Add some utilities for testing network
Added a sketch and a simple server written in Ruby to debug network
issues. If these two things work, your problem is elsewhere.

Given that I'm still seeing issues with sending temperature reading
updates to Elasticsearch, I must have an issue with my HTTP requests or
Elasticsearch or something.

Interesting to note: judging by watching the timing on the logs, it
appears that IP address resolution is far and away the slowest part of
the request-sending process...