"Fixed" stalling bug
This is one of the more infuriating bugs I've had to deal with in a
while. The code kept stalling in the `client.fastrprint()` function, and
it was completely unclear why. After lots of debug messages and
commenting out even remotely-related code, it was commenting out the
"Content-Type: application/json" line that made it all start working.

Completely and utterly baffled.

Now to actually send the data, plus clean up the mess.